add sorting in correct places
